Move Home, CharacterListPage and SingleCharacterPage to subfolder pages
