+++ src/pages/CharacterListPage.css
