+++ src/pages/SingleCharacterPage.css
